+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/exception.cpp
@@ -24,7 +24,7 @@ CommandExceptionHelp()
 
     ShowMessages(
         "syntax : \t!exception [IdtIndex (hex)] [pid ProcessId (hex)] "
-        "[core CoreId (hex)] [imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+        "[core CoreId (hex)] [imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
         "[script { Script (string) }] [condition { Condition (hex) }] [code { Code (hex) }]\n");
 
     ShowMessages("\nnote: monitoring page-faults (entry 0xe) is implemented differently.\n");