+++ hyperdbg/hprdbgctrl/code/debugger/commands/extension-commands/tsc.cpp
@@ -22,7 +22,7 @@ CommandTscHelp()
     ShowMessages("!tsc : monitors execution of rdtsc/rdtscp instructions.\n\n");
 
     ShowMessages("syntax : \t!tsc [pid ProcessId (hex)] [core CoreId (hex)] "
-                 "[imm IsImmediate (yesno)] [buffer PreAllocatedBuffer (hex)] "
+                 "[imm IsImmediate (yesno)] [sc EnableShortCircuiting (onoff)] [buffer PreAllocatedBuffer (hex)] "
                  "[script { Script (string) }] [condition { Condition (hex) }] "
                  "[code { Code (hex) }]\n");